+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Update `GetBitmapFrame` to round `width` that is changed by `dpi`.
 - Add `GetBitmapFrameByWidthAndDpi` to get optimal frame by `dpi` and `width`.
 - Add `GetSystemDpi` to get the system `dpi` on the fly.
+- Add `SystemDpi` to store the system `dpi` value.
+- Update `GetBitmapFrameByWidthAndDpi` to round `dpi` frame value.
 ### Example
 - Add `Cube-Grey-Light.tiff` and `Cube-Grey-Dark.tiff` in `AppTheme`.